Fix a bunch of issues

+++ code/pilotfile/csg.cpp
@@ -480,7 +480,7 @@ void pilotfile::csg_write_techroom()
 	startSection(Section::Techroom);
 
 	// visible ships
-    for (int i = 0; i < Ship_info.size(); ++i) {
+    for (size_t i = 0; i < Ship_info.size(); ++i) {
         if ((Ship_info[i].flags[Ship::Info_Flags::In_tech_database]) && !(Ship_info[i].flags[Ship::Info_Flags::Default_in_tech_database])) {
             visible = 1;
         }
